+++ src/database/src/Commands/ModelCommand.php
@@ -93,7 +93,8 @@ class ModelCommand extends Command
             ->setPrefix($this->getOption('prefix', 'prefix', $pool, ''))
             ->setInheritance($this->getOption('inheritance', 'commands.db:model.inheritance', $pool, 'Model'))
             ->setUses($this->getOption('uses', 'commands.db:model.uses', $pool, 'Hyperf\DbConnection\Model\Model'))
-            ->setForceCasts($this->getOption('force-casts', 'commands.db:model.force_casts', $pool, false));
+            ->setForceCasts($this->getOption('force-casts', 'commands.db:model.force_casts', $pool, false))
+            ->setRefreshFillable($this->getOption('refresh-fillable', 'commands.db:model.refresh_fillable', $pool, false));
 
         if ($table) {
             $this->createModel($table, $option);
@@ -112,6 +113,7 @@ class ModelCommand extends Command
         $this->addOption('prefix', 'P', InputOption::VALUE_OPTIONAL, 'What prefix that you want the Model set.');
         $this->addOption('inheritance', 'i', InputOption::VALUE_OPTIONAL, 'The inheritance that you want the Model extends.');
         $this->addOption('uses', 'U', InputOption::VALUE_OPTIONAL, 'The default class uses of the Model.');
+        $this->addOption('refresh-fillable', null, InputOption::VALUE_NONE, 'Whether generate fillable argement for model.');
     }
 
     protected function getSchemaBuilder(string $poolName): MySqlBuilder
@@ -158,7 +160,10 @@ class ModelCommand extends Command
 
         $stms = $this->astParser->parse(file_get_contents($path));
         $traverser = new NodeTraverser();
-        $visitor = make(ModelUpdateVisitor::class, ['columns' => $columns]);
+        $visitor = make(ModelUpdateVisitor::class, [
+            'columns' => $columns,
+            'option' => $option,
+        ]);
         $traverser->addVisitor($visitor);
         $stms = $traverser->traverse($stms);
         $code = $this->printer->prettyPrintFile($stms);
@@ -203,7 +208,7 @@ class ModelCommand extends Command
     protected function getOption(string $name, string $key, string $pool = 'default', $default = null)
     {
         $result = $this->input->getOption($name);
-        $nonInput = $name === 'force-casts' ? false : null;
+        $nonInput = in_array($name, ['force-casts', 'refresh-fillable']) ? false : null;
         if ($result === $nonInput) {
             $result = $this->config->get("databases.{$pool}.{$key}", $default);
         }